banner completed

+++ build_erp_backend/src/application/UseCase/Estimation/UpdateEstimation.ts
@@ -8,19 +8,23 @@ import { IUpdateEstimationUseCase } from '../../IUseCases/IEstimation/IUpdateEst
 
 export class UpdateEstimationUsecase implements IUpdateEstimationUseCase {
 
-    constructor(
-      private _estimationRepository: IEstimationRepository,
-      private _stageRepository: IStageRepository,
-    ) { }
-    async execute(input: { projectId: string, row: saveEstimationInput[] }):
-      Promise<commonOutput> {
-        const { projectId, row } = input;
-        const existStage = await this._stageRepository.findStageByprojectId(projectId);
-        if (existStage.length > 0) {
-            return ResponseHelper.conflictData(EstimationFailedMessage.USED_STAGE);
-        }
-        await this._estimationRepository.deleteEstimationsByProjectId(projectId);
-        await this._estimationRepository.createEstimation(row, projectId);
-        return ResponseHelper.success(EstimationSuccessMessage.UPDATE);
+  constructor(
+    private _estimationRepository: IEstimationRepository,
+    private _stageRepository: IStageRepository,
+  ) { }
+  async execute(input: { projectId: string, row: saveEstimationInput[] }):
+    Promise<commonOutput> {
+    const { projectId, row } = input;
+    const existEstimation = await this._estimationRepository.getEstimationsByProjectId(projectId)
+    if (existEstimation) {
+      return ResponseHelper.conflictData(EstimationFailedMessage.ALREADY_DONE)
     }
+    const existStage = await this._stageRepository.findStageByprojectId(projectId);
+    if (existStage.length > 0) {
+      return ResponseHelper.conflictData(EstimationFailedMessage.USED_STAGE);
+    }
+    await this._estimationRepository.deleteEstimationsByProjectId(projectId);
+    await this._estimationRepository.createEstimation(row, projectId);
+    return ResponseHelper.success(EstimationSuccessMessage.UPDATE);
+  }
 }
