+++ build_erp_backend/src/application/UseCase/SiteManager/DeleteSitemanager.ts
@@ -1,5 +1,6 @@
+import { IprojectRepository } from '../../../domain/Entities/IRepository/IProject';
 import { ISitemanagerRepository } from '../../../domain/Entities/IRepository/ISitemanager';
-import { SitemanagerSuccessMessage } from '../../../Shared/Messages/Sitemanager.Message';
+import { SitemanagerFailedMessage, SitemanagerSuccessMessage } from '../../../Shared/Messages/Sitemanager.Message';
 import { ResponseHelper } from '../../../Shared/responseHelpers/response';
 import { commonOutput } from '../../dto/common';
 import { IDeleteSitemanagerUseCase } from '../../IUseCases/ISitemanager/IDeleteSitemanager';
@@ -8,12 +9,17 @@ import { IDeleteSitemanagerUseCase } from '../../IUseCases/ISitemanager/IDeleteS
 
 
 export class DeleteSitemanagerUseCase implements IDeleteSitemanagerUseCase {
-    constructor(
-      private _sitemanagerRepository: ISitemanagerRepository,
-    ) { }
-    async execute(id: string): Promise<commonOutput> {
-        await this._sitemanagerRepository.removeSitemanagerById(id);
-        return ResponseHelper.success(SitemanagerSuccessMessage.DELETE);
+  constructor(
+    private _sitemanagerRepository: ISitemanagerRepository,
+    private _projectRepository: IprojectRepository
+  ) { }
+  async execute(id: string): Promise<commonOutput> {
+    const existSitemanagerInProject =  await this._projectRepository.getProjectsBySitemanagerId(id)
+    if(existSitemanagerInProject){
+      return ResponseHelper.conflictData(SitemanagerFailedMessage.ADD_PROJECT)
     }
+    await this._sitemanagerRepository.removeSitemanagerById(id);
+    return ResponseHelper.success(SitemanagerSuccessMessage.DELETE);
+  }
 
 }
